update to work with new Gerrit and add customization (#6)
update to work with new Gerrit and add customization

This commit updates the script to move away from its jQuery usage, allowing
it to load on "old new Gerrit". It also detects "new new Gerrit" (PolyGerrit),
to be compatible with it.

It also adds the ability to customize the checklist per repo, along with
customizing the default value. The default value customization allows those
who want to enforce stricter adherence to the checklist to default the value
to `no` instead of `N/A`. This means if a reviewer doesn't check any boxes,
it will show that they didn't check them rather than not showing anything.